Compilation fix: Explict casting on logtwo function.
git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@443 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/include/PrismHybridGlob.h
@@ -83,7 +83,7 @@ extern double last_unif;
 
 // macros, function prototypes
 
-#define logtwo(X) log(X)/log(2)
+#define logtwo(X) log((double)X)/log(2.0)
 void PH_PrintToMainLog(JNIEnv *env, char *str, ...);
 void PH_PrintToTechLog(JNIEnv *env, char *str, ...);
 void PH_SetErrorMessage(char *str, ...);
